refactor: remove components barrel file
