+++ components/Common/Form/MultiSelectInput/MultiSelectInput.tsx
@@ -1,6 +1,6 @@
 import { useRef, useState, FC } from 'react';
 import { KeyboardEventHandler } from 'react';
-import { messageType, toastMessage } from '@/components';
+import { messageType, toastMessage } from '@/components/Toaster';
 import toast from 'react-hot-toast';
 import classNames from 'classnames';
