--- components/index.ts
@@ -1,12 +1,0 @@
-export * from './AuthModal';
-export * from './Common/Button';
-export * from './ErrorFallback';
-export * from './Common/Form';
-export * from './LandingPage';
-export * from './ProductHunt';
-export * from './ProjectsPage/ProjectsList/ProjectsList';
-export * from './AddProjectForm';
-export * from './Tags';
-export * from './Common/Typography';
-export * from './Toaster';
-export * from './Layouts';
