+++ pages/posts/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type { NextPage } from 'next';
-import { SharedLayout } from '@/components';
+import { SharedLayout } from '@/components/Layouts';
 
 const Posts: NextPage = () => {
   return (
